contract api refactor - cleanup
