+++ elrond-wasm/src/contract_base/contract_base_trait.rs
