+++ elrond-wasm/src/contract_base/mod.rs
@@ -1,12 +1,12 @@
 mod contract_abi_provider;
-mod contract_base;
+mod contract_base_trait;
 mod contract_traits;
 mod proxy_obj_base;
 mod proxy_obj_callback_base;
 mod wrappers;
 
 pub use contract_abi_provider::*;
-pub use contract_base::*;
+pub use contract_base_trait::ContractBase;
 pub use contract_traits::*;
 pub use proxy_obj_base::ProxyObjApi;
 pub use proxy_obj_callback_base::CallbackProxyObjApi;
